compute the ambiguity cache for each language, as part of define-language
but only after verifying that for every define-language in the examples
and in the benchmark, it takes less than 10 msec to compute it except
on the r6rs grammar, where it takes 75 msec. The 75 msec is almost all spent
in the n^2 production overlap checking. There are 42 non-terminals
and 168 productions, which seems pretty big and not too too bad timewise
so I'll leave it always being computed for now

also, move the ambiguity test suite into redex/tests